refactor(test_tools): move get_test_cases to the span loader as setup_test_cases
OkahuEval.get_test_cases was mostly trace and span gathering, not
evaluation, so it and its _fact_spans helper move to OkahuSpanLoader and
call the loader's own methods directly. The eval report stays with the
evaluator; setup_test_cases reaches it through a local import, since
okahu_eval imports this module.

The public fluent get_test_cases(source=...) entry point is unchanged and
now dispatches to the new location.
